[TFX-2240] future-value-for-eod-thre

See merge request tfx-clearing/eod-service!93

+++ src/main/java/com/ihsmarkit/tfx/eod/service/EODCalculator.java
@@ -349,7 +349,7 @@ public class EODCalculator {
     }
 
     public Map<CurrencyPairEntity, List<BalanceTrade>> rebalanceLPPositions(final Stream<ParticipantPositionEntity> positions,
-                                                                            final Map<CurrencyPairEntity, BigDecimal> thresholds) {
+                                                                            final Map<CurrencyPairEntity, Long> thresholds) {
 
         return positions
             .map(tradeOrPositionMapper::convertPosition)
